feat: update resource manager
