--- packages/runtime/resource/.npmignore
@@ -1,5 +1,0 @@
-src/
-node_modules/
-__tests__/
-api-extractor.json
-.npmignore
