--- packages/runtime/resource/README.md
@@ -1,11 +1,0 @@
-# `@garfish/resource`
-
-> TODO: description
-
-## Usage
-
-```
-import @garfish/resource from '@garfish/resource';
-
-// TODO: DEMONSTRATE API
-```
