--- packages/runtime/resource/__tests__/index.spec.ts
@@ -1,5 +1,0 @@
-import toolApp from '../src/index';
-
-describe('resource', () => {
-  it('needs tests');
-});
