--- packages/runtime/resource/src/component.ts
@@ -1,1 +1,0 @@
-export class ComponentResource {}
