--- packages/runtime/resource/src/style.ts
@@ -1,1 +1,0 @@
-export class StyleResource {}
