+++ packages/runtime/utils/src/index.ts
@@ -6,3 +6,9 @@ export * from './container';
 export * from './webpack';
 export * from './sentry';
 export * from './garfish';
+
+// resource manager
+export * from './resourceManager/style';
+export * from './resourceManager/component';
+export * from './resourceManager/javascript';
+export * from './resourceManager/template/main';
