+++ packages/runtime/utils/src/resourceManager/README.md
@@ -1,0 +1,15 @@
+# Process request resource
+
+- template resource
+
+  - parse
+  - render
+
+- css resource
+
+  - transform style
+  - transform url
+  - add style scope
+
+- javascript resource
+  - unknow
