+++ packages/runtime/utils/src/resourceManager/component.ts
@@ -1,0 +1,1 @@
+export class ComponentManager {}
